+++ custom_components/kd_brain/heatpump/__init__.py
@@ -1,0 +1,1 @@
+"""Heat pump optimization (setpoint-offset shifting, M6b)."""